requirements.txt
httpx
playwright
beautifulsoup4
markdownify
pydantic
tenacity
python-dotenv
openai
pandas
aiosqlite
tqdm
streamlit
